+++ NEWS.md
@@ -43,6 +43,7 @@
 * The Data Viewer now uses the `format()` methods defined for columns entries when available (#7239)
 * Add support for navigating source history with mouse forward/back buttons (#7272)
 * Improved error logging of mistyped usernames when using PAM authentication (#7501)
+* Add ability to go directly to various Global Option panes via Command Palette (#7678)
 * R6Class method defintions are now indexed and accessible by the fuzzy finder (Ctrl + .)
 
 ### RStudio Server Pro